+++ core/src/main/scala/chisel3/experimental/verification/package.scala
@@ -1,0 +1,41 @@
+// See LICENSE for license details.
+
+package chisel3.experimental
+
+import chisel3.{Bool, CompileOptions}
+import chisel3.internal.Builder
+import chisel3.internal.Builder.pushCommand
+import chisel3.internal.firrtl.{Formal, Verification}
+import chisel3.internal.sourceinfo.SourceInfo
+
+package object verification {
+  object assert { // scalastyle:ignore object.name
+    def apply(predicate: Bool, msg: String = "")(
+      implicit sourceInfo: SourceInfo,
+      compileOptions: CompileOptions): Unit = {
+      val clock = Builder.forcedClock
+      pushCommand(Verification(Formal.Assert, sourceInfo, clock.ref,
+        predicate.ref, msg))
+    }
+  }
+
+  object assume { // scalastyle:ignore object.name
+    def apply(predicate: Bool, msg: String = "")(
+      implicit sourceInfo: SourceInfo,
+      compileOptions: CompileOptions): Unit = {
+      val clock = Builder.forcedClock
+      pushCommand(Verification(Formal.Assume, sourceInfo, clock.ref,
+        predicate.ref, msg))
+    }
+  }
+
+  object cover { // scalastyle:ignore object.name
+    def apply(predicate: Bool, msg: String = "")(
+      implicit sourceInfo: SourceInfo,
+      compileOptions: CompileOptions): Unit = {
+      val clock = Builder.forcedClock
+      pushCommand(Verification(Formal.Cover, sourceInfo, clock.ref,
+        predicate.ref, msg))
+    }
+  }
+}
